pagiantion-working

+++ src/main/java/com/rainbowtape/boards/controller/ForumController.java
@@ -1,13 +1,17 @@
 package com.rainbowtape.boards.controller;
 
 import java.security.Principal;
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
@@ -18,6 +22,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.rainbowtape.boards.entity.Post;
@@ -32,10 +37,10 @@ import com.rainbowtape.boards.service.UserService;
 @Controller
 @RequestMapping(value = "/forum")
 public class ForumController {
-
+	
 	@Autowired
 	private PostService postService;
-	
+
 	@Autowired
 	private ReplyService replyService;
 
@@ -49,7 +54,7 @@ public class ForumController {
 	public Post makePostAttribute () {
 		return new Post();
 	}
-	
+
 	@ModelAttribute("reply")
 	public Reply makeReplyAttribute () {
 		return new Reply();
@@ -81,24 +86,32 @@ public class ForumController {
 		if (result.hasErrors()) {
 			System.out.println(result.toString());
 			// 플래쉬어트리뷰트를 이용하면, model 어트리뷰트로 추가해 주지 않아도, 리다이렉트된 콘트롤러에서 한번만 사용되고 자동으로 지워진다. 
-			redirectAttributes.addFlashAttribute("errormsg", "'글종류'를 포함한, 제목과 내용을 모두 작성해 주셔야 합니다. HTML 코드입력은 허용되지않습니다. 문자로만 입력하여 작성해 주세요."); 
+			redirectAttributes.addFlashAttribute("errormsg", "입력에 실패하였습니다. 정상적으로 다시 작성해 주세요."); 
+//			redirectAttributes.addAttribute("post", post);
 			return "redirect:/forum/write";
 		}
 
 		post.setUser(user);
 		java.util.Date now = new java.util.Date();
 		post.setDatecreated(now);
+		post.setDatemodified(now);
 
 		postService.save(post);
 		return "redirect:/forum/viewpostslist";
 	}
 
 	@GetMapping("/viewpostslist")
-	public String showPostsList (@ModelAttribute("post") Post post, Model model) {
-
-		List<Post> posts = new ArrayList<Post>();
-		posts = postService.getAllPost();
-		model.addAttribute("posts", posts);
+	public String showPostsList (
+			@ModelAttribute("post") Post post,
+			@PageableDefault(page = 1) Pageable pageable,
+			Model model) {
+		
+		System.err.println(pageable.getPageNumber());
+		int pageNum = (pageable.getPageNumber() == 0) ? 0 : (pageable.getPageNumber() - 1);
+		pageable = new PageRequest(pageNum, 10, new Sort(Sort.Direction.DESC, "datemodified"));
+		
+		Page<Post> page = postService.findAll(pageable);
+		model.addAttribute("page", page);
 
 		return "_viewpostslist";
 	}
@@ -107,70 +120,71 @@ public class ForumController {
 	public String showPost (
 			@ModelAttribute("post") Post post, 
 			@ModelAttribute("reply") Reply reply,
-			@PathVariable int idpost, Model model, 
-			@ModelAttribute("user") User user, 
+			@PathVariable int idpost, Model model,
 			HttpServletRequest request) {
 
 		post = postService.findById(idpost);
 		List<Reply> replys = replyService.findByPost(post);
-		
-//		boolean isThisUserOriginalPosterOrAdmin = false;
-//		User originalPoster = post.getUser();
-//		if (user.getId() == originalPoster.getId() || request.isUserInRole("ROLE_ADMIN")){
-//			isThisUserOriginalPosterOrAdmin = true;
-//		}
-		
+
 		model.addAttribute("post", post);
 		model.addAttribute("replys", replys);
-		//model.addAttribute("isThisUserOriginalPosterOrAdmin", isThisUserOriginalPosterOrAdmin);
 
 		return "_viewpost";
 	}
 
-	@PreAuthorize("#user.email == authentication.name")
+	@PreAuthorize("#userid == #user.id")
 	@GetMapping("/update/{idpost}")
 	public String showUpdateForm(
 			@PathVariable int idpost, 
-			@ModelAttribute("post") Post post, 
+			@ModelAttribute("post") Post post,
 			@ModelAttribute("user") User user,
+			@RequestParam("u") int userid,
 			Model model) {
 
 		post = postService.findById(idpost);
 		model.addAttribute("post", post);
 		return "_updateForm";
 	}
-	
-	@PreAuthorize("#user.email == authentication.name")
+
 	@PostMapping("/update/{idpost}")
 	public String updatePost(
 			@PathVariable int idpost, 
 			@ModelAttribute("post") @Valid Post temp, 
 			@ModelAttribute("user") User user,
 			BindingResult result) {
-		
+
 		Post originalPost = postService.findById(idpost);
-		
+
 		originalPost.setCategory(temp.getCategory());
 		originalPost.setTitle(temp.getTitle());
 		originalPost.setContent(temp.getContent());
 		originalPost.setTag(temp.getTag());
-		
+
 		java.util.Date now = new java.util.Date();
 		originalPost.setDatemodified(now);
-		
 		postService.update(originalPost);
-		
-		return "redirect:/forum/viewpostslist";
+
+		return "redirect:/forum/viewpost/" + idpost;
 	}
-	
-	@PreAuthorize("#user.email == authentication.name or hasRole('ROLE_ADMIN')")
+
+	@PreAuthorize("#userid == #user.id or hasRole('ROLE_ADMIN')")
 	@PostMapping("/delete/{idpost}")
-	public String deletePost (@PathVariable int idpost, @ModelAttribute("user") User user) {
+	public String deletePost (
+			@PathVariable int idpost, 
+			@ModelAttribute("user") User user,
+			@RequestParam("u") int userid,
+			RedirectAttributes redirectAttributes) {
+
+		if (!postService.findById(idpost).getReplys().isEmpty()) {
+			System.err.println(postService.findById(idpost).getReplys().size());
+			redirectAttributes.addFlashAttribute("errormsg", "댓글이 달린 글은 삭제할수 없습니다."); 
+			return "redirect:/forum/viewpost/" + idpost;
+		}
 
 		postService.delete(idpost);
 		return "redirect:/forum/viewpostslist";
 	}
-	
+
 	@PostMapping("/writeReply/{idpost}")
 	public String storeReplyIntoDb (
 			@ModelAttribute("reply") @Valid Reply reply, 
@@ -185,21 +199,28 @@ public class ForumController {
 			redirectAttributes.addFlashAttribute("errormsg", "댓글의 내용을 작성해 주셔야 합니다. html 태그는 허용되지 않습니다."); 
 			return "redirect:/forum/viewpost/" + idpost;
 		}
-		
+
 		reply.setUser(user);
 		java.util.Date now = new java.util.Date();
 		reply.setDatecreated(now);
-		
+
 		Post post = postService.findById(idpost);
+		post.setDatemodified(now);
+		System.err.println(post.getDatemodified());
+		postService.save(post);
+
 		reply.setPost(post);
-		
 		replyService.save(reply);
 
 		return "redirect:/forum/viewpost/" + idpost;
 	}
-	
-	
-	
-	
+
+	@PostMapping("/deleteReply/{idreply}")
+	public String deleteReply (@PathVariable int idreply, @RequestParam("idpost") int idpost, @ModelAttribute("user") User user) {
+
+		replyService.delete(idreply);
+
+		return "redirect:/forum/viewpost/" + idpost;
+	}
 
 }
